start retrieving contacts from sqlite database
Currently, ABEnumContacts() function has an array hard coded with some
fake entries. We did this way to be able to develop, tet and debug the
other features while we were working in a way to retrieve the records from
the database but not slow down the overall development.

Now, the code in the Linux code base reached a point where it can be
ported to the Windows side and I'm satisfied with.

To get it started, I'm adding the rows retrieved from the contacts table
to the array with the fake entries and returning that to the caller to
fill the list view.

I'll be keeping the fake entries until I make sure the "New",
"Properties", and "Delete" features are working properly, are fully tested
and debugged.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>