tribe: fix variable naming issues in ab_init()
Currently, in ab_init() function, there is this confusion with these
three variables. dbpath, dbfile and dbname are not named using snake_case
convention, but they should. So, in this change we fix that.

Now, we also rename this variables for clarity, to make the code easier
to understand. dbpath was renamed to db_dir to clarify the caller should
pass a path to the directory where the database file will be saved.
Similarily, dbfile was renamed to db_file_path to be clear that it is the
final concatenation of the directory and file name. Finally, dbname was
renamed to db_file_name to make it clear it is the file name of the database,
because sqlite does not have the concept of "CREATE DATABASE name". The
database is in the file.

We used the A/HC/LC pattern, which in this case "db" was the prefix rather
than the action, the main context (high context) was "file", and the low
context was used to help narrow the concept.

Refs:
https://johnfarrier.com/a-practical-guide-for-naming-things/
https://medium.com/@code.ceeker/naming-conventions-camel-case-pascal-case-kebab-case-and-more-dc4e515b9652

Signed-off-by: Thiago Farina <tfransosi@gmail.com>