tribe: remove redundant sentence from schema.sql
The sentence is stating the obvious and it is not
adding any relevant or new information. What it is
saying can be inferred from the path and all the
context.

It is also written in ALL UPPER CASE, which comments
like that should not be written. We should normally
write text in "Sentence style".

Signed-off-by: Thiago Farina <tfransosi@gmail.com>