remove arraysize.h file
The Win32 API has an ARRAYSIZE macro defined in winnt.h file and this
header file normally gets included when including windows.h.

There, for C it is implemented in terms of RTL_NUMBER_OF_V1, which
does the same calculation of our ARRAYSIZE macro.

So, on Windows, there is no much point in providing our own version of
this macro in a dedicated header file. Unless, we were unifying the code
base with other platforms like Linux and Mac. But for now, that is not
in my plans.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>