tribe: make use of GNUInstallDirs module
In the past we introduced BINDIR variable at commit af8c6bc69dc1
("tribe: define BINDIR variable and use it) and DATADIR variable
at 9bb72703660c ("tribe: define a desktopdir variable and use it")
to provide standard GNU installation directories as defined by GNU
Coding Standards [0].

The problem is, this is complicated, there are many details that
should be addressed and is very subtle.

CMake's GNUInstallDirs module [1] does all the work for us and probably
gets all the details right.

So, instead of reinventing the wheel and doing a poor job like we were
doing, let's reuse this module, which is probably battle tested by
many projects.

One last note, the cmake code from Inkscape was really helpful on this.

Refs:
[0] - https://www.gnu.org/prep/standards/html_node/Directory-Variables.html
[1] - https://github.com/Kitware/CMake/blob/master/Modules/GNUInstallDirs.cmake

Signed-off-by: Thiago Farina <tfransosi@gmail.com>