tribe: remove original programs
Why? Because I don't have time and resources (money) to keep
working and maintaining two fronts. The code in the
standalone programs are exactly the same as the ones in the
cmd_xxx. So, everytime I make a change to a function in the
api, I have to replicate the changes in two places.

With the introduction of the command line tool at
commit 3294641d3a4e ("Add cli version.") they became basically
obsolete and almost a burden, as now there are two places to
patch instead of just one.

Below are the relevant commits I was able find:

Commit 26cbe0b815d6 ("Add create_users_table.c") introduced
the users table that we still use today but with "contacts"
name.

Commit 6bc85bfe9f72 ("Add select_from_user_table.c") formed
the basis for the "ls" command.

Commit 110b0ef17419 ("Add insert_into_user_table.c") formed
the basis for the "add" command.

Commit 7a88fe242a1d ("Add update_user_table.c") formed the
basis for the "mod" command.

Commit dda9e0270004 ("Add a new prototype of delete.") added
a new "del" command.

And finally, commit 3294641d3a4e ("Add cli version.") added
the command line tool that unified all these standalone
programs into a single binary driver.

That leaves us with tribecli and tribe executables to
concentrate, focus and work on.

The good thing with version control systems is that, we can
bring them back, in the future, if necessary.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>