tribe: pack scrolled window before creating the list view
Previously, we were creating the list view, then attaching it to the scrolled
window and at the end packing the scrolled window.

Now, looking more closely, it seems to make more sense to create the scrolled
window first, pack it into the vbox, then create the list view, and finally
add it to the scrolled window.

I didn't notice any visual difference, so it looks like it is correct.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>