tribe: check glib dependency
GLib is a required dependency because GTK depends on it.

This version 2.19.7 was chosen because gtk 2.16.1 depended
on it.

See https://gitlab.gnome.org/GNOME/gtk/-/blob/035be761f4df2d41bb1684bba77cb87758af9c05/configure.in#L34.

GLib 2.19.7 was released on February 16, 2009. 15 years ago. It was
declared unstable development release. So, not very good in terms
of stability and we better move both libraries to versions that
are stable releases. https://lwn.net/Articles/319725/

The oldest tag/version in
https://gitlab.gnome.org/GNOME/glib/-/tags?sort=version_asc is
2.20.0 from March 26, 2010.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>