tribe: remove extraneous whitespace in PACKAGE variable definition
This was added with this much whitespace between the variable name
and the value at commit aff1d5276c2f ("sab: add 'dist' target to
create a tarball from the source files"), because it was copied from
network/echod/CMakeLists.txt, and there the PACKAGE_XXX values were
vertical aligned.

But here we don't we align vertically, so let's remove the extra
whitespace.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>