tribe: rename VERSION_EXTRA to VERSION_SUFFIX
It was originally "EXTRA", because of the projects I looked
at. So, why change it to "SUFFIX"? Because, I think it explains
better what this component of the version string is. It is
a suffix because the appended string can be an "-alpha",
"-beta", "-dev", "-rc1", depending on the development cycle.

And "extra" looks like "additional" to me, but it is not an
additional version.

Some projects, like Inkspace and odpi use "suffix". Bind and
mysql use "extra". So there are both factions.

So, let's rename this macro to VERSION_SUFFIX.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>